Less GC pressure if I'm not allocating a new struct, map, and array for each and every request.

The trade off is I have to iterate over the map and `delete` every entry from it before it can be reused.

Also updated the benchmarks to use `ServeHTTP` now that the code path is different between it and `Handler`