@chrisaxiom 2017-05-12
are there any benchmarks before/after?