@AndrewBurian 2017-05-12
Ugh. Benchmarks aren't super great. 

Before
```
BenchmarkSingleRoute-8      	 2000000	       692 ns/op	     528 B/op	       9 allocs/op
BenchmarkShallowAndWide-8   	 2000000	       960 ns/op	     528 B/op	       9 allocs/op
BenchmarkNarrowAndDeep-8    	  100000	     14699 ns/op	    2288 B/op	       9 allocs/op
BenchmarkFan-8              	 2000000	       941 ns/op	     528 B/op	       9 allocs/op
PASS
ok  	github.com/andrewburian/powermux	9.715s
 ```
After
```
benchmem -bench .
BenchmarkSingleRoute-8      	 2000000	       729 ns/op	     480 B/op	       8 allocs/op
BenchmarkShallowAndWide-8   	 1000000	      1009 ns/op	     480 B/op	       8 allocs/op
BenchmarkNarrowAndDeep-8    	  100000	     14632 ns/op	    2240 B/op	       8 allocs/op
BenchmarkFan-8              	 1000000	      1004 ns/op	     480 B/op	       8 allocs/op
```

Doesn't look like the allocs saved and GC time saved outweighs the cost of running the sync pool