@AndrewBurian 2017-05-12
Oh shii... didn't realize how much overhead there was on a defer call.

New After:
```
BenchmarkSingleRoute-8      	 2000000	       651 ns/op	     480 B/op	       8 allocs/op
BenchmarkShallowAndWide-8   	 2000000	       939 ns/op	     480 B/op	       8 allocs/op
BenchmarkNarrowAndDeep-8    	  100000	     14353 ns/op	    2240 B/op	       8 allocs/op
BenchmarkFan-8              	 2000000	       905 ns/op	     480 B/op	       8 allocs/op
PASS
ok  	github.com/andrewburian/powermux	9.411s
```