@acroca 2017-05-12
Performance wise the pool doesn't seem to help much, it'll for sure reduce some gc time and keep the memory less fragmented, but I don't think the introduced complexity (and risk of not doing a proper clean up when putting things back to the pool) justifies the feature.