@AndrewBurian 2017-05-13
Drove it down a bit more reusing `pathParts` slices. Saved a lot of memory on the worst case routes, but somehow still no fewer allocs.

```
BenchmarkSingleRoute-8      	 3000000	       633 ns/op	     480 B/op	       8 allocs/op
BenchmarkShallowAndWide-8   	 2000000	       866 ns/op	     480 B/op	       8 allocs/op
BenchmarkNarrowAndDeep-8    	  200000	     10581 ns/op	     481 B/op	       8 allocs/op
BenchmarkFan-8              	 2000000	       798 ns/op	     480 B/op	       8 allocs/op
PASS
ok  	github.com/andrewburian/powermux	10.047s
```