@AndrewBurian 2017-05-13
A few more tricks and we're now saving far more substantially on allocs and run time

```
BenchmarkSingleRoute-8      	 3000000	       443 ns/op	     400 B/op	       5 allocs/op
BenchmarkShallowAndWide-8   	 2000000	       717 ns/op	     400 B/op	       5 allocs/op
BenchmarkNarrowAndDeep-8    	  200000	     10120 ns/op	     401 B/op	       5 allocs/op
BenchmarkFan-8              	 2000000	       645 ns/op	     400 B/op	       5 allocs/op
PASS
ok  	github.com/andrewburian/powermux	8.305s
```