@turbolent 2017-05-14
I'm with @acroca here: introducing pools adds quite some complexity compared to very little performance gain, making the library harder to understand and maintain. 

Nice job though on the other refactoring and small optimizations, like duplicate `EscapedURL` calls.

I'd keep it simple and not have the pools, but 👍 for the optimizations.