@AndrewBurian 2017-05-14
I tested it with just the pools taken out, and it adds between 10% and 16% back on to the runtime. It also disproportionately hurts the kind of routes that could really use the most help, which are the ones with long paths or high number of path parameters. 14000ns down to 10000ns for the long path is quite an improvement.

In terms of code readability, the two pool `Get`/`Put` blocks are each contained to a single short function so you can at a glance see it coming and going. Those functions don't ever handle errors or return early so there's a low risk of missing a `Put` later. The `routeExecution` struct needed a constructor anyways to DRY up the few places it was being created, so adding the pool around it and doing it all in one file doesn't make it too much more confusing in my opinion.