@AndrewBurian 2017-05-15
Rewrote the history to take out the benchmark rewrite since that has been cherry picked to master.
You can now directly compare the performance on master to the performance on this branch.