@AndrewBurian 2017-05-15
It's been a long time since I got my teeth into a good optimization project. And of all the components of web service, this is the one I'd argue the most for milking every cycle out of it.

Also blame @Kay-Zee for the cherry pick 😛