@turbolent 2017-05-15
OK, thanks for the clarification Andrew. Code looks clean, so let's get this merged