@Kay-Zee 2017-05-15
While I agree that readability is important, I think that the code is still quite clean, and a 30% increase is negligible. 

We've also done a lot to circumvent possible mistakes, and check for race-conditions, so if the savings really is 30%, then I think its worth keeping.